Apply feedbacks
